+++ game-server/src/com/xeno/entity/actor/player/Equipment.java
@@ -54,6 +54,10 @@ public class Equipment {
 				return false;
 			} 
 		}
+		
+		if (!p.getMapZoneManager().execute(p, zone -> zone.canEquip(p, slot, itemID)))
+			return false;
+		
 		if (!p.getInventory().deleteItem(itemID, slot, amount)) {
 			return false;
 		}
